# backend/openapi.json
{
    "openapi":  "3.1.0",
    "info":  {
                 "title":  "AcademiaConnect API",
                 "description":  "Backend API for AcademiaConnect SIH 2026",
                 "version":  "1.0.0"
             },
    "paths":  {
...
                  "/student/profile/{user_id}":  {
                                                     "get":  {
                                                                 "tags":  [
                                                                              "Student Profile"
                                                                          ],
                                                                 "summary":  "Get Student Profile",
                                                                 "operationId":  "get_student_profile_student_profile__user_id__get",
                                                                 "parameters":  [
                                                                                    {
                                                                                        "name":  "user_id",
                                                                                        "in":  "path",
                                                                                        "required":  true,
                                                                                        "schema":  {
                                                                                                       "type":  "integer",
                                                                                                       "title":  "User Id"
                                                                                                   }
                                                                                    }
                                                                                ],
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "Successful Response",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {

                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           },
                                                                                   "422":  {
                                                                                               "description":  "Validation Error",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {
                                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           }
                                                                               }
                                                             }
                                                 },
...
                  "/company/profile/{user_id}":  {
                                                     "post":  {
                                                                  "tags":  [
                                                                               "Company"
                                                                           ],
                                                                  "summary":  "Create Company Profile",
                                                                  "operationId":  "create_company_profile_company_profile__user_id__post",
                                                                  "parameters":  [
                                                                                     {
                                                                                         "name":  "user_id",
                                                                                         "in":  "path",
                                                                                         "required":  true,
                                                                                         "schema":  {
                                                                                                        "type":  "integer",
                                                                                                        "title":  "User Id"
                                                                                                    }
                                                                                     }
                                                                                 ],
                                                                  "requestBody":  {
                                                                                      "required":  true,
                                                                                      "content":  {
                                                                                                      "application/json":  {
                                                                                                                               "schema":  {
                                                                                                                                              "$ref":  "#/components/schemas/CompanyProfileCreate"
                                                                                                                                          }
                                                                                                                           }
                                                                                                  }
                                                                                  },
                                                                  "responses":  {
                                                                                    "200":  {
                                                                                                "description":  "Successful Response",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {

                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            },
                                                                                    "422":  {
                                                                                                "description":  "Validation Error",
                                                                                                "content":  {
                                                                                                                "application/json":  {
                                                                                                                                         "schema":  {
                                                                                                                                                        "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                    }
                                                                                                                                     }
                                                                                                            }
                                                                                            }
                                                                                }
                                                              },
                                                     "get":  {
                                                                 "tags":  [
                                                                              "Company Profile"
                                                                          ],
                                                                 "summary":  "Get Company Profile",
                                                                 "operationId":  "get_company_profile_company_profile__user_id__get",
                                                                 "parameters":  [
                                                                                    {
                                                                                        "name":  "user_id",
                                                                                        "in":  "path",
                                                                                        "required":  true,
                                                                                        "schema":  {
                                                                                                       "type":  "integer",
                                                                                                       "title":  "User Id"
                                                                                                   }
                                                                                    }
                                                                                ],
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "Successful Response",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {

                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           },
                                                                                   "422":  {
                                                                                               "description":  "Validation Error",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {
                                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           }
                                                                               }
                                                             },
                                                     "put":  {
                                                                 "tags":  [
                                                                              "Company"
                                                                          ],
                                                                 "summary":  "Update Company Profile",
                                                                 "operationId":  "update_company_profile_company_profile__user_id__put",
                                                                 "parameters":  [
                                                                                    {
                                                                                        "name":  "user_id",
                                                                                        "in":  "path",
                                                                                        "required":  true,
                                                                                        "schema":  {
                                                                                                       "type":  "integer",
                                                                                                       "title":  "User Id"
                                                                                                   }
                                                                                    }
                                                                                ],
                                                                 "requestBody":  {
                                                                                     "required":  true,
                                                                                     "content":  {
                                                                                                     "application/json":  {
                                                                                                                              "schema":  {
                                                                                                                                             "$ref":  "#/components/schemas/CompanyProfileCreate"
                                                                                                                                         }
                                                                                                                          }
                                                                                                 }
                                                                                 },
                                                                 "responses":  {
                                                                                   "200":  {
                                                                                               "description":  "Successful Response",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {

                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           },
                                                                                   "422":  {
                                                                                               "description":  "Validation Error",
                                                                                               "content":  {
                                                                                                               "application/json":  {
                                                                                                                                        "schema":  {
                                                                                                                                                       "$ref":  "#/components/schemas/HTTPValidationError"
                                                                                                                                                   }
                                                                                                                                    }
                                                                                                           }
                                                                                           }
                                                                               }
                                                             }
                                                 },
...
              },
    "components":  {
                       "schemas":  {
...
                                       "CompanyProfileCreate":  {
                                                                    "properties":  {
                                                                                       "name":  {
                                                                                                    "type":  "string",
                                                                                                    "title":  "Name"
                                                                                                },
                                                                                       "industry":  {
                                                                                                        "type":  "string",
                                                                                                        "title":  "Industry",
                                                                                                        "default":  ""
                                                                                                    },
                                                                                       "description":  {
                                                                                                           "type":  "string",
                                                                                                           "title":  "Description",
                                                                                                           "default":  ""
                                                                                                       },
                                                                                       "location":  {
                                                                                                        "type":  "string",
                                                                                                        "title":  "Location",
                                                                                                        "default":  ""
                                                                                                    }
                                                                                   },
                                                                    "type":  "object",
                                                                    "required":  [
                                                                                     "name"
                                                                                 ],
                                                                    "title":  "CompanyProfileCreate"
                                                                },
...
                                       "HTTPValidationError":  {
                                                                   "properties":  {
                                                                                      "detail":  {
                                                                                                     "items":  {
                                                                                                                   "$ref":  "#/components/schemas/ValidationError"
                                                                                                               },
                                                                                                     "type":  "array",
                                                                                                     "title":  "Detail"
                                                                                                 }
                                                                                  },
                                                                   "type":  "object",
                                                                   "title":  "HTTPValidationError"
                                                               },
...
                                       "ValidationError":  {
                                                               "properties":  {
                                                                                  "loc":  {
                                                                                              "items":  {
                                                                                                            "anyOf":  [
                                                                                                                          {
                                                                                                                              "type":  "string"
                                                                                                                          },
                                                                                                                          {
                                                                                                                              "type":  "integer"
                                                                                                                          }
                                                                                                                      ]
                                                                                                        },
                                                                                              "type":  "array",
                                                                                              "title":  "Location"
                                                                                          },
                                                                                  "msg":  {
                                                                                              "type":  "string",
                                                                                              "title":  "Message"
                                                                                          },
                                                                                  "type":  {
                                                                                               "type":  "string",
                                                                                               "title":  "Error Type"
                                                                                           }
                                                                              },
                                                               "type":  "object",
                                                               "required":  [
                                                                                "loc",
                                                                                "msg",
                                                                                "type"
                                                                            ],
                                                               "title":  "ValidationError"
                                                           },
...
                                   }
                   }
}